+++ tests/test_within_intercept.R
@@ -77,7 +77,7 @@ f_level_gi_u <- fixef(gi_u, type = "level")
 f_dmean_gi_u <- fixef(gi_u, type = "dmean")
 
 
-# in the one-way unbalanced case: is the overall intercept is the weighted mean of the effects
+# in the one-way unbalanced case: is the overall intercept is the _weighted_ mean of the effects
 # (with the current fixef implementation) - this check also depends on how type = "dmean" is calculated in fixef
 int_gi_u <- within_intercept(gi_u)
 mod_int_gi_u <- within_intercept(gi_u, return.model = TRUE)
@@ -85,16 +85,10 @@ individual_intercepts_gi_u <- int_gi_u + f_dmean_gi_u
 
 int_manual_gi_u <- weighted.mean(fixef(gi_u), as.numeric(table(index(gi_u)[[1]])))
 mean(f_level_gi_u)
-individual_intercepts_gi_u2 <- mean(f_level_gi_u) + f_dmean_gi_u
 
 # check consistency of functions in themselves
-# if (!isTRUE(all.equal(individual_intercepts_gi_u,  f_level_gi_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
-# if (!isTRUE(all.equal(individual_intercepts_gi_u2, f_level_gi_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
-
-#if (!isTRUE(all.equal(int_gi_u, int_manual_gi_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
-
-# take weighted.mean:
-# if (!isTRUE(all.equal(individual_intercepts_gi_u2, f_level_gi_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
+if (!isTRUE(all.equal(individual_intercepts_gi_u,  f_level_gi_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
+if (!isTRUE(all.equal(int_gi_u, int_manual_gi_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
 
 
 # oneway time unbalanced
@@ -105,27 +99,16 @@ int_gt_u <- within_intercept(gt_u)
 mod_int_gt_u <- within_intercept(gt_u, return.model = TRUE)
 individual_intercepts_gt_u <- int_gt_u + f_dmean_gt_u
 
+int_manual_gt_u <- weighted.mean(fixef(gt_u), as.numeric(table(index(gt_u)[[2]])))
+mean(f_level_gt_u) # mean is not correct for unbalanced case!
+int_gt_u <- within_intercept(gt_u)
 
-# fails for unbalanced models, likely due to fixef calc. for dmean: fixef - mean(fixef); need weighted mean?
+
 # check consistency of functions in themselves
-# if (!isTRUE(all.equal(individual_intercepts_gt_u, f_level_gt_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
+if (!isTRUE(all.equal(individual_intercepts_gt_u, f_level_gt_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
+if (!isTRUE(all.equal(int_gt_u, int_manual_gt_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
 
-## take weighted mean for "dmean" calc. instead:
-# int_manual_gt_u <- weighted.mean(fixef(gt_u), as.numeric(table(index(gt_u)[[2]])))
-# mean(f_level_gt_u)
-# 
-# int_gt_u <- within_intercept(gt_u)
-# 
-# if (!isTRUE(all.equal(int_gt_u, int_manual_gt_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
-# 
-# f_dmean_gt_u2 <- f_level_gt_u - int_manual_gt_u
-# individual_intercepts_gt_u2 <- int_gt_u + f_dmean_gt_u2
-# # taking weighted.mean: works!
-# if (!isTRUE(all.equal(individual_intercepts_gt_u2, f_level_gt_u, check.attributes = FALSE))) stop("within_intercept, unbalanced: something is wrong")
-# 
-
-
-
+ 
 ## twoways unbalanced
 gtw_u <- plm(inv ~ value + capital, data = Grunfeld_unbalanced, model = "within", effect = "twoways")
 f_level_tw_i_u <- fixef(gtw_u, type = "level", effect = "individual")
@@ -135,18 +118,19 @@ f_dmean_tw_t_u <- fixef(gtw_u, type = "dmean", effect = "time")
 
 int_tw_u <- within_intercept(gtw_u)
 
-weights_gtw_i_u <- as.numeric(table(index(gtw_u)[[1]]))
-weights_gtw_t_u <- as.numeric(table(index(gtw_u)[[2]]))
+weights_gtw_i_u <- as.numeric(table(index(gtw_u)[[1L]]))
+weights_gtw_t_u <- as.numeric(table(index(gtw_u)[[2L]]))
 
-# in the unbalanced case, the means of the level fixed effects are not the same
-# (like they were in the balanced case)
+# in the unbalanced twoway case, the means of the level fixed effects are not the same
+# (like they were in the balanced twoway case)
 int_manual_tw_i_u <- mean(f_level_tw_i_u)
 int_manual_tw_t_u <- mean(f_level_tw_t_u)
 all.equal(int_manual_tw_i_u, int_manual_tw_t_u) # not equal
 # ... but weighted means are
-int_manual_tw_i_u2 <- as.vector(crossprod(f_level_tw_i_u, weights_gtw_i_u) / sum(weights_gtw_i_u))
-int_manual_tw_t_u2 <- as.vector(crossprod(f_level_tw_t_u, weights_gtw_t_u) / sum(weights_gtw_t_u))
+int_manual_tw_i_u2 <- weighted.mean(f_level_tw_i_u, w = pdim(gtw_u)$Tint$Ti)
+int_manual_tw_t_u2 <- weighted.mean(f_level_tw_t_u, w = pdim(gtw_u)$Tint$nt)
 
+int_manual_tw_u <- int_manual_tw_i_u2 # any of individual or time will do
 
 individual_intercepts_tw_i_u <- int_tw_u + f_dmean_tw_i_u
 individual_intercepts_tw_t_u <- int_tw_u + f_dmean_tw_t_u
@@ -157,15 +141,11 @@ individual_intercepts_tw_t_u2 <- int_manual_tw_t_u2 + f_dmean_tw_t_u
 mod_lm <- lm(inv ~ value + capital + factor(firm) + factor(year), data = Grunfeld_unbalanced)
 
 # check consistency of functions fixef and within_intercept
-# fails for unbalanced models, likely due to fixef calc. for dmean: fixef - mean(fixef); need weighted mean?
-
-#if (!isTRUE(all.equal(individual_intercepts_tw_i_u2, f_level_tw_i_u, check.attributes = FALSE))) stop("within_intercept twoways, individual: something is wrong")
-#if (!isTRUE(all.equal(individual_intercepts_tw_t_u2, f_level_tw_t_u, check.attributes = FALSE))) stop("within_intercept twoways, time:       something is wrong")
-#if (!isTRUE(all.equal(individual_intercepts_tw_i_u, f_level_tw_i_u, check.attributes = FALSE))) stop("within_intercept twoways, individual: something is wrong")
-#if (!isTRUE(all.equal(individual_intercepts_tw_t_u, f_level_tw_t_u, check.attributes = FALSE))) stop("within_intercept twoways, time:       something is wrong")
-#if (!isTRUE(all.equal(int_tw_u, int_manual_tw_u, check.attributes = FALSE))) stop("within_intercept: something is wrong")
-
-
+if (!isTRUE(all.equal(individual_intercepts_tw_i_u2, f_level_tw_i_u, check.attributes = FALSE))) stop("within_intercept twoways, individual: something is wrong")
+if (!isTRUE(all.equal(individual_intercepts_tw_t_u2, f_level_tw_t_u, check.attributes = FALSE))) stop("within_intercept twoways, time:       something is wrong")
+if (!isTRUE(all.equal(individual_intercepts_tw_i_u, f_level_tw_i_u, check.attributes = FALSE))) stop("within_intercept twoways, individual: something is wrong")
+if (!isTRUE(all.equal(individual_intercepts_tw_t_u, f_level_tw_t_u, check.attributes = FALSE))) stop("within_intercept twoways, time:       something is wrong")
+if (!isTRUE(all.equal(int_tw_u, int_manual_tw_u, check.attributes = FALSE))) stop("within_intercept: something is wrong")
 
 
 ### print all within intercepts (to have them compared to the reference output test_within_intercept.Rout.save)