+++ NEWS.md
@@ -43,6 +43,8 @@ plm()):
   (such as lag, diff, nobs, ...), so that the help systems offers to access the
   plm-specific documentation (regression introduced when pkg plm 2.0-0 adopted
   roxygen2 for documentation).
+* fixef(x, type = "dmean"): for unbalanced data, deviation from weighted mean is
+  calculated.
 * ercomp: cosmetic: if one of theta\$id, theta\$time is 0 => theta\$total must be 
   0 and is set to 0 (before, for some data and platforms, theta$total could be a
   very small positive or negative number, due to limited computational precision).